Extracts configspace node to separate type
